+++ src/components/Slider/Caddyfile
@@ -1,0 +1,35 @@
+{
+    email server@mstagency.ru
+}
+www.sbcart.mstprime.ru {
+    redir https://sbcart.mstprime.ru{uri}
+}
+www.sbdelivery.mstprime.ru {
+    redir https://sbdelivery.mstprime.ru{uri}
+}
+www.sbguest.mstprime.ru {
+    redir https://sbguest.mstprime.ru{uri}
+}
+www.sbgeoservice.mstprime.ru {
+    redir https://sbgeoservice.mstprime.ru{uri}
+}
+sbcart.mstprime.ru {
+	reverse_proxy nginx:80 {
+		header_up Host sbcart.mstprime.ru
+	}
+}
+sbdelivery.mstprime.ru {
+	reverse_proxy nginx:80 {
+		header_up Host sbdelivery.mstprime.ru
+	}
+}
+sbguest.mstprime.ru {
+	reverse_proxy nginx:80 {
+		header_up Host sbguest.mstprime.ru
+	}
+}
+sbgeoservice.mstprime.ru {
+	reverse_proxy nginx:80 {
+		header_up Host sbgeoservice.mstprime.ru
+	}
+}
