+++ src/components/Slider/__stories__/Slider.stories.tsx
@@ -1,4 +1,6 @@
-import * as React from 'react';
+import './Slider.stories.css';
+
+import React, { useEffect, useState } from 'react';
 import { action } from '@storybook/addon-actions';
 import { boolean, number, select, text } from '@storybook/addon-knobs';
 
@@ -9,16 +11,21 @@ import {
   eventInterceptorMap,
   EventInterceptorProvider,
 } from '../../EventInterceptor/EventInterceptor';
-import { TextField } from '../../TextField/TextField';
-import { Slider, sliderPropSize, sliderPropSizeDefault } from '../Slider';
+import {
+  TextField,
+  TextFieldPropOnChange,
+  textFieldPropSize,
+  textFieldPropSizeDefault,
+} from '../../TextField/TextField';
+import { Slider } from '../Slider';
 
 import mdx from './Slider.docs.mdx';
 
 const defaultKnobs = () => ({
   disabled: boolean('disabled', false),
   division: boolean('division', true),
-  size: select('Size', sliderPropSize, sliderPropSizeDefault),
   classname: text('classname', 'inputRange'),
+  size: select('size', textFieldPropSize, textFieldPropSizeDefault),
   step: number('step', 25),
   customStep: boolean('customStep', false),
   min: number('minValue', -50),
@@ -32,25 +39,59 @@ const defaultKnobs = () => ({
 const cnSliderStories = cn('SliderStories');
 
 export function Playground() {
-  const { range, prefix, suffix, customStep, size, step, ...props } = defaultKnobs();
+  const {
+    range,
+    prefix,
+    suffix,
+    disabled,
+    customStep,
+    size,
+    step,
+    min,
+    max,
+    ...props
+  } = defaultKnobs();
+  const [value, setValue] = useState([50]);
 
-  const value = React.useMemo(() => (range ? [0, 60] : 0), [range]);
+  const handleChange: TextFieldPropOnChange = (e) =>
+    setValue((prev) => (prev.length > 1 ? [Number(e.value), prev[1]] : [Number(e.value)]));
+
+  useEffect(() => {
+    setValue(range ? [0, 50] : [50]);
+  }, [range]);
 
   return (
     <EventInterceptorProvider eventHandler={console.log} map={eventInterceptorMap}>
       <div className={cnSliderStories()}>
         <Slider
+          disabled={disabled}
+          min={min}
+          max={max}
           {...props}
           prefix={
             prefix &&
             (({ value }) => (
-              <TextField size={size} value={(Array.isArray(value) ? value[0] : value).toString()} />
+              <TextField
+                type="number"
+                size={size}
+                min={min}
+                max={max}
+                step={step}
+                value={(Array.isArray(value) ? value[0] : value || 0).toString()}
+                className={cnSliderStories('textfield', { prefix })}
+                onChange={handleChange}
+                disabled={disabled || range}
+                required
+              />
             ))
           }
-          suffix={suffix && <IconSettings />}
+          suffix={
+            suffix && <IconSettings className={cnSliderStories('icon', { disabled, suffix })} />
+          }
           step={customStep ? [20, 17, 22, 20] : step}
-          value={value}
+          value={range ? value : value[0]}
           onChange={action('onChange')}
+          onChangeCommitted={(_, v) => setValue(v)}
         />
       </div>
     </EventInterceptorProvider>
