+++ src/components/Slider/sbcart.mstprime.ru.conf
@@ -1,0 +1,85 @@
+# Redirect every request to HTTPS...
+#server {
+#    listen 80;
+#    listen [::]:80;
+
+#    server_name .sbcart.mstprime.ru;
+#    return 301 https://$host$request_uri;
+#}
+
+
+# Redirect SSL to primary domain SSL...
+#server {
+    #listen 443 ssl http2;
+    #listen [::]:443 ssl http2;
+    
+    # FORGE SSL (DO NOT REMOVE!)
+    #ssl_certificate /etc/nginx/ssl/sbcart.sushibox.org/976524/server.crt;
+    #ssl_certificate_key /etc/nginx/ssl/sbcart.sushibox.org/976524/server.key;
+
+    #ssl_protocols TLSv1.2 TLSv1.3;
+    #ssl_ciphers TLS13-AES-256-GCM-SHA384:TLS13-CHACHA20-POLY1305-SHA256:TLS_AES_256_GCM_SHA384:TLS-AES-256-GCM-SHA384:TLS_CHACHA20_POLY1305_SHA256:TLS-CHACHA20-POLY1305-SHA256:ECDHE-ECDSA-AES256-GCM-SHA384:ECDHE-ECDSA-CHACHA20-POLY1305:ECDHE-ECDSA-AES256-SHA384:ECDHE-RSA-AES256-GCM-SHA384:ECDHE-RSA-CHACHA20-POLY1305:ECDHE-RSA-AES256-SHA384:ECDHE-ECDSA-AES256-SHA:ECDHE-RSA-AES256-SHA;
+    #ssl_prefer_server_ciphers on;
+    #ssl_dhparam /etc/nginx/dhparams.pem;
+
+    #server_name www.sbcart.mstprime.ru;
+    #return 301 https://sbcart.mstprime.ru$request_uri;
+#}
+
+server {
+    #listen 443 ssl http2;
+    #listen [::]:443 ssl http2;
+	listen 80;
+    #listen [::]:80;
+    server_name sbcart.mstprime.ru;
+    server_tokens off;
+    root /var/www/sbcart/public;
+
+    # FORGE SSL (DO NOT REMOVE!)
+    #ssl_certificate /etc/nginx/ssl/sbcart.sushibox.org/976524/server.crt;
+    #ssl_certificate_key /etc/nginx/ssl/sbcart.sushibox.org/976524/server.key;
+
+    #ssl_protocols TLSv1.2 TLSv1.3;
+    #ssl_ciphers TLS13-AES-256-GCM-SHA384:TLS13-CHACHA20-POLY1305-SHA256:TLS_AES_256_GCM_SHA384:TLS-AES-256-GCM-SHA384:TLS_CHACHA20_POLY1305_SHA256:TLS-CHACHA20-POLY1305-SHA256:ECDHE-ECDSA-AES256-GCM-SHA384:ECDHE-ECDSA-CHACHA20-POLY1305:ECDHE-ECDSA-AES256-SHA384:ECDHE-RSA-AES256-GCM-SHA384:ECDHE-RSA-CHACHA20-POLY1305:ECDHE-RSA-AES256-SHA384:ECDHE-ECDSA-AES256-SHA:ECDHE-RSA-AES256-SHA;
+    #ssl_prefer_server_ciphers on;
+    #ssl_dhparam /etc/nginx/dhparams.pem;
+
+    #add_header X-Frame-Options "SAMEORIGIN";
+    #add_header X-XSS-Protection "1; mode=block";
+    #add_header X-Content-Type-Options "nosniff";
+
+    index index.html index.htm index.php;
+
+    charset utf-8;
+
+    #location /.well-known/acme-challenge {
+	#	auth_basic off;
+	#	allow all;
+	#	alias /home/forge/.letsencrypt;
+	#}
+
+    location / {
+        try_files $uri $uri/ /index.php?$query_string;
+    }
+
+    location = /favicon.ico { access_log off; log_not_found off; }
+    location = /robots.txt  { access_log off; log_not_found off; }
+
+    access_log off;
+    error_log  /var/log/nginx/sbcart.mstprime.ru-error.log error;
+
+    error_page 404 /index.php;
+
+    location ~ \.php$ {
+        #fastcgi_split_path_info ^(.+\.php)(/.+)$;
+		fastcgi_param  SCRIPT_FILENAME  $document_root$fastcgi_script_name;
+        fastcgi_pass sbcart:9000;
+        fastcgi_index index.php;
+        include fastcgi_params;
+    }
+
+    location ~ /\.(?!well-known).* {
+        deny all;
+    }
+}
+
